chore: remove ConfigBase and StaticConfigBase classes
Remove ConfigBase and StaticConfigBase classes to simplify the usage of both `$staticConfig` and `$config` - we do not need the base classes - just move the implementation to the current inheritors
